Explain 4GB RAM requirement for setup computers
The setup computers will be running Ubuntu live, and need to download
the Ubuntu ISO to write to the Boot USBs. In a live system, the
writable filesystems are all RAM disks. The Ubuntu ISO is 1.6GB.

With only 2GB RAM, I got about halfway through the download before it
aborted with "No space left on device."

Fixes issue #25.